Refactoring: split interpretation logic for Hyper and Tableau query plans
So far, loading Hyper and Tableau query plans shared some logic such as
the `generateDisplayNames` function. This function ended up to handle
both plan types at the same time and hence it was hard to change, e.g.,
the way Hyper plans were loaded without accidentally breaking
functionality for Tableau plans. This commit separates the loading logic
for both of them.
